Add pid controlled impedance for each joint
We switch the control approach to using a PID position controller around
an effort interface and superimpose the previous effort command as a
feedforward torque.  An additional feedback linearization for gravity
and overall dynamics will be added in later commits.

This will allow users to control the robot with controllers from both
the `effort_controllers` and `position_controllers` families.
In the case of the `position_controllers`, each joint impedance (spring
damper behavior) can be adjusted separately via dynamic reconfigure.